[portsorch] Avoid the situation where the bridge port of VLAN memeber cannot be removed after VLAN removal
In some cases, when vlan memeber has been remove, but bridge port is not removed beause that fdb_count is not zero.
During the retry, the vlan is removed, and then the retry of removeBridgePort will not be executed.
Therefore, In delete task of vlan member, it is not need to check vlan is existed.