FDB flush and make sure the FDB count is 0 before removing VLAN member
If FDB ageout notifications occurs after the removal of a member port from a VLAN.
It becomes impossible to locate the correct physical port through the flow port, resulting in the inability to notify swss about the age-out event.
This issue arises due to the disassociation between the flow port and the physical port.

The solution involves preemptively flushing the member ports of the VLAN before removing the member from the VLAN.
It's essential to confirm the successful deletion of swss member port FDB entries.
Only after this confirmation should the member be removed from the VLAN.