[portsorch] Check FDB count is cleared before removing bridge port
Ensured that the FDB count is checked and cleared before allowing the bridge port to be removed.

The bridge port should not be removed if there are associated FDB entries, meaning the FDB count is greater than 0.
Removing the bridge port before clearing the FDB entries leads to FDB update failures.

The following steps is able to reproduce the issue.
1. Remove a VLAN member with associated FDB entries.
2. The bridge port would be removed without checking the FDB count is cleared.
3. Receive FDB event for the bridge port, then failed to update the FDB.
4. Remove VLAN is failed because the FDB count of the VLAN is not cleared
5. Keep retrying to remove the VLAN