remove old fdb entry from map during MOVE event
Fdb entry consist of mac, vlan and port_name fields. It seems that
during seeking through map there are only mac and vlan is taken into
account. So when mac move occured, new entry with updated port_name
is not placed in the map, only just changed fdbdata that leads to
incorrect port m_fdb_count decrement in the future.

The proposal is to remove old entry from map when mac move occured.
This guarantee placement of new entry in the map.

Signed-off-by: Yurii Tretiakov <ytretiakov@larch-networks.com>
Signed-off-by: Anton Parkhomenko <aparkhomenko@larch-networks.com>